% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_odata_get.R
\name{get_catalogue}
\alias{get_catalogue}
\title{GET the catalogue of Stats NZ OData API.}
\usage{
get_catalogue(timeout = 60)
}
\arguments{
\item{timeout}{Timeout for the GET request, in seconds.}
}
\value{
A data frame containing the requested catalogue data. This is
generally not a "tidy dataset" since it uses prefixes to represent some of
the more nested content.
}
\description{
Derived from https://github.com/StatisticsNZ/open-data-api/
}
