% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_secrets.R
\name{set_secrets}
\alias{set_secrets}
\title{Set app secrets}
\usage{
set_secrets(...)
}
\arguments{
\item{...}{Key-value pairs to be saved as "secrets" (all arguments should
be named).}
}
\description{
Sets secret for app, to be accessed by app during run-time or
by other exported functions. Saves it to global options, so this is stored
in memory for the lifetime of the current R session.

P.S. don't have secrets hard-coded in your own code, read them in from
somewhere secretive
}
