% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsnz.odata.R
\docType{package}
\name{statsnz.odata-package}
\alias{statsnz.odata}
\alias{statsnz.odata-package}
\title{statsnz.odata: Simple app to browse Stats OData API}
\description{
Graphical interface to browse the datasets in the Stats OData API and get help with building queries.
}
\details{
Stats NZ OData API Client
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/xaviermiles/statsnz.odata}
  \item Report bugs at \url{https://github.com/xaviermiles/statsnz.odata/issues}
}

}
\author{
\strong{Maintainer}: Xavier Miles \email{xaviermiles377@gmail.com}

}
\keyword{internal}
