% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internals.R
\docType{package}
\name{statsnz.odata.client-package}
\alias{statsnz.odata.client}
\alias{statsnz.odata.client-package}
\title{statsnz.odata.client: Simple app to browse Stats OData API}
\description{
Graphical interface to browse the datasets in the Stats Odata API and get help with building queries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/xaviermiles/statsnz.odata.client}
  \item Report bugs at \url{https://github.com/xaviermiles/statsnz.odata.client/issues}
}

}
\author{
\strong{Maintainer}: Xavier Miles \email{xaviermiles377@gmail.com}

}
\keyword{internal}
