% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-odata.R
\name{tbl_odata}
\alias{tbl_odata}
\title{Create an odata tbl (abstract?)}
\usage{
tbl_odata(subclass, endpoint, entity, ..., vars = NULL)
}
\arguments{
\item{subclass}{Name of subclass.}

\item{vars}{Column names as a character vector}
}
\description{
Create an odata tbl (abstract?)
}
