% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_odata_get.R
\name{basic_get}
\alias{basic_get}
\title{basic_get}
\usage{
basic_get(endpoint, entity, query = "", timeout = 60)
}
\arguments{
\item{endpoint}{API endpoint. Required.}

\item{entity}{Data entity. Required.}

\item{query}{Query URL character (not URL-encoded).}

\item{timeout}{Timeout for the GET request(s), in seconds.}
}
\value{
A data frame containing the requested data.
}
\description{
Basic GET request to Stats NZ OData API.

Derived from https://github.com/StatisticsNZ/open-data-api/
}
