% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_odata_get.R
\name{get_catalogue}
\alias{get_catalogue}
\title{GET the catalogue of Stats NZ OData API.}
\usage{
get_catalogue(timeout = 10)
}
\arguments{
\item{timeout}{Timeout for the GET request, in seconds.}
}
\value{
A data frame containing the requested catalogue data.
}
\description{
Derived from https://github.com/StatisticsNZ/open-data-api/

The returned data is generally not a "tidy dataset" so the R data frame
will use prefixes to represent the nested content.
}
